Resolve the issue of clinky timeouts
The problem: the timeout responsible for recursive call of newGame
function had to be inside of the timeout containing 'prompt' in order
to not be executed before the input has been provided. It turns out
the pause in code execution caused by 'prompt' affects only the local
scope, or, perhaps, it works this way only with setTimeout.